requirements.txt
fastapi[standard]==0.113.0
pymysql==1.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
